+++ CHANGELOG.md
@@ -8,7 +8,8 @@ ADDED:
  
 - new factory methods for LinkedHashSet and InsertionOrderSet created
 - new hashcode method for collections that contains array objects
-- new equalCollections method for collections created
+- new isEqualCollection method for collections created
+- new method for copy an array created
 
 CHANGED: